// Client/src/entities/orders/model/order-status.ts
export const enum OrderStatus {
	Pending = "Pending",
	Confirmed = "Confirmed", //
	Processing = "Processing", //
	ReadyForShipment = "ReadyForShipment", //
	Shipped = "Shipped", //
	OutForDelivery = "OutForDelivery", //
	Delivered = "Delivered", //
	Completed = "Completed", //
	Cancelled = "Cancelled", //
	Returned = "Returned", //
	Failed = "Failed", //
	UnderReview = "UnderReview" //
}